AST/CommentSema.cpp: Fix signess in abs() to appease msvc. It would not make sense to pass (unsigned)-(unsigned) to abs().
llvm-svn: 160097

+++ clang/lib/AST/CommentSema.cpp
@@ -390,7 +390,8 @@ unsigned Sema::correctTypoInParmVarReference(
     const IdentifierInfo *II = ParamVars[i]->getIdentifier();
     if (II) {
       StringRef Name = II->getName();
-      unsigned MinPossibleEditDistance = abs(Name.size() - Typo.size());
+      unsigned MinPossibleEditDistance =
+        abs((int)Name.size() - (int)Typo.size());
       if (MinPossibleEditDistance > 0 &&
           Typo.size() / MinPossibleEditDistance < 3)
         continue;
